4. move inner classes to upper level
