+++ src/test/java/com/together/buytogether/member/feature/SEX.java
@@ -1,0 +1,12 @@
+package com.together.buytogether.member.feature;
+
+public enum SEX {
+    MALE("남자"),
+    FEMALE("여자");
+
+    private final String description;
+
+    SEX(String description) {
+        this.description = description;
+    }
+}
