--- src/test/java/com/together/buytogether/service/RegisterMemberTest.java
@@ -1,181 +1,0 @@
-package com.together.buytogether.service;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.util.Assert;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-
-public class RegisterMemberTest {
-    private RegisterMember registerMember;
-    private MemberRepository memberRepository;
-
-    @BeforeEach
-    void setUp() {
-        memberRepository = new MemberRepository();
-        registerMember = new RegisterMember(memberRepository);
-    }
-
-    @Test
-    @DisplayName("회원가입")
-    void registerMember() {
-        String address = "경기도 고양시 덕양구 화정로 27";
-        String detailAddress = "625동 1004호";
-        RegisterMember.Request request = new RegisterMember.Request(
-                "name",
-                "loginId",
-                "password",
-                "phoneNumber",
-                SEX.MALE,
-                address, //도로명 주소
-                detailAddress    //상세 주소
-        );
-        //when
-        registerMember.request(request);
-
-        //then
-        assertThat(memberRepository.findAll().size()).isEqualTo(1);
-    }
-
-    public enum SEX {
-        MALE("남자"),
-        FEMALE("여자");
-
-        private final String description;
-
-        SEX(String description) {
-            this.description = description;
-        }
-    }
-
-    public static class Address {
-        private final String address;
-        private final String detailAddress;
-
-        public Address(String address, String detailAddress) {
-            validateAddress(address, detailAddress);
-            this.address = address;
-            this.detailAddress = detailAddress;
-        }
-
-        private void validateAddress(String address, String detailAddress) {
-            Assert.hasText(address, "주소는 필수 값입니다");
-            Assert.hasText(detailAddress, "상세 주소는 필수 값입니다");
-        }
-    }
-
-    public static class Member {
-        private Long memberId;
-        private final String name;
-        private final String loginId;
-        private final String password;
-        private final String phoneNumber;
-        private final SEX sex;
-        private final Address address;
-
-        public Member(
-                String name,
-                String loginId,
-                String password,
-                String phoneNumber,
-                SEX sex,
-                Address address) {
-            validateMember(name, loginId, password, phoneNumber, sex, address);
-            this.name = name;
-            this.loginId = loginId;
-            this.password = password;
-            this.phoneNumber = phoneNumber;
-            this.sex = sex;
-            this.address = address;
-        }
-
-        private void validateMember(
-                String name,
-                String loginId,
-                String password,
-                String phoneNumber,
-                SEX sex,
-                Address address) {
-            Assert.hasText(name, "이름은 필수 값입니다");
-            Assert.hasText(loginId, "로그인 아이디는 필수 값입니다");
-            Assert.hasText(password, "비밀번호는 필수 값입니다");
-            Assert.hasText(phoneNumber, "전화번호는 필수 값입니다");
-            Assert.notNull(sex, "성별은 필수 값입니다");
-            Assert.notNull(address, "주소는 필수 값입니다");
-        }
-
-        public void assignId(Long memberId) {
-            this.memberId = memberId;
-        }
-
-        public Long getId() {
-            return memberId;
-        }
-    }
-
-    private static class MemberRepository {
-        private final Map<Long, Member> members = new HashMap<>();
-        private Long memberId = 0L;
-
-        public void save(Member member) {
-            member.assignId(memberId++);
-            members.put(member.getId(), member);
-        }
-
-        public List<Member> findAll() {
-            return new ArrayList<>(members.values());
-        }
-    }
-
-    private class RegisterMember {
-        private final MemberRepository memberRepository;
-
-
-        public RegisterMember(MemberRepository memberRepository) {
-            this.memberRepository = memberRepository;
-        }
-
-        public void request(Request request) {
-            // request에서 필요한 값들을 꺼내서 회원 도메인을 생성하고 저장한
-            Member member = request.toDomain();
-            memberRepository.save(member);
-        }
-
-        public record Request(
-                String name,
-                String loginId,
-                String password,
-                String phoneNumber,
-                SEX sex,
-                String address,
-                String detailAddress) {
-            public Request {
-                Assert.hasText(name, "이름은 필수 값입니다");
-                Assert.hasText(loginId, "로그인 아이디는 필수 값입니다");
-                Assert.hasText(password, "비밀번호는 필수 값입니다");
-                Assert.hasText(phoneNumber, "전화번호는 필수 값입니다");
-                Assert.notNull(sex, "성별은 필수 값입니다");
-                Assert.hasText(address, "주소는 필수 값입니다");
-                Assert.hasText(detailAddress, "상세 주소는 필수 값입니다");
-            }
-
-            public Member toDomain() {
-                return new Member(
-                        name,
-                        loginId,
-                        "password",
-                        "phoneNumber",
-                        sex,
-                        new Address(address, detailAddress)
-                );
-            }
-        }
-    }
-}
